# Meta election perf