This address 3 different issue that occur when having many events and that are blocking generating benchmarks for multiple 1000's of events.

1- is_already_interesting_content doing linear search on events.
2- On consensus, avoid expensive vote resolving if not the needed payload key.
3- Avoid N^2 on number of unconsensused events in num_creators_of_ancestors_carrying_payload


Combined performance Test results

time cargo bench --features=testing -- --baseline pr267 t1024

a_node4_opaque_evt1024 - bench_section_size_evt1024_interleave
                        time:   [101.69 ms 101.74 ms 101.85 ms]
                        change: [-68.700% -68.632% -68.576%] (p = 0.02 < 0.05)
                        Performance has improved.
a_node8_opaque_evt1024 - bench_section_size_evt1024_interleave
                        time:   [293.68 ms 317.17 ms 328.20 ms]
                        change: [-71.050% -69.783% -67.682%] (p = 0.02 < 0.05)
                        Performance has improved.#
a_node16_opaque_evt1024 - bench_section_size_evt1024_interleave
                        time:   [706.60 ms 710.02 ms 710.42 ms]
                        change: [-56.815% -56.624% -56.444%] (p = 0.02 < 0.05)
                        Performance has improved.
a_node32_opaque_evt1024 - bench_section_size_evt1024_interleave
                        time:   [2.3159 s 2.3210 s 2.3313 s]
                        change: [-45.645% -45.469% -45.298%] (p = 0.02 < 0.05)
                        Performance has improved.